requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.15
django-allauth==0.47.0
django-ckeditor==6.4.2
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-pandas==0.6.6
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-summernote==0.8.20.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
dynaconf==3.1.7
et-xmlfile==1.1.0
filelock==3.4.2
greenlet==1.1.2
gunicorn==20.1.0
hvac==0.11.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
knox==0.1.14
loguru==0.5.3
MarkupSafe==2.0.1
numpy==1.23.1
oauthlib==3.1.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.0.1
platformdirs==2.4.1
psycopg2==2.9.2
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
virtualenv==20.13.0
webencodings==0.5.1
whitenoise==5.3.0
